Made One Class -- ListUtils, with 2 static methods(Loop, NonLoop). Update DataManagerTest: try-with-resources.
