+++ src/test/java/com/github/olegik1719/trello/DataManagerTest.java
@@ -15,15 +15,20 @@ public class DataManagerTest {
     public void getInstance() throws PropertyVetoException, SQLException {
         ComboPooledDataSource h2 = new DataManager().getInstance();
         assertTrue("h2 is null!",h2!=null);
-        Connection conn = h2.getConnection();
-        Statement statement = conn.createStatement();
-        statement.execute("CREATE TABLE Persons (PersonID int,LastName varchar(255),FirstName varchar(255),"
-                 +" Address varchar(255), City varchar(255))");
-        statement.execute("INSERT INTO Persons VALUES(1, 'HELLO', 'HELLO','HELLO','HELLO')");
-        statement.execute("INSERT INTO Persons VALUES(2, 'JOHN','JOHN','JOHN','JOHN')");
-        ResultSet resultSet = statement.executeQuery("SELECT * FROM Persons");
-        while (resultSet.next()){
-            System.out.printf("%s%n", resultSet.getString(2));
+        try(Connection conn = h2.getConnection();
+            Statement statement = conn.createStatement()
+        ){
+            statement.execute("CREATE TABLE Persons (PersonID int,LastName varchar(255),FirstName varchar(255),"
+                    + " Address varchar(255), City varchar(255))");
+            statement.execute("INSERT INTO Persons VALUES(1, 'HELLO', 'HELLO','HELLO','HELLO')");
+            statement.execute("INSERT INTO Persons VALUES(2, 'JOHN','JOHN','JOHN','JOHN')");
+            try (ResultSet resultSet = statement.executeQuery("SELECT * FROM Persons")
+            ) {
+                while (resultSet.next()) {
+                    System.out.printf("%s%n", resultSet.getString(2));
+                }
+            }
         }
+
     }
 }
